refactor: domain 폴더 생성
