+++ src/domain/LottoSeller.js
