+++ src/domain/PrizeCalculator.js
