+++ src/domain/RankCalculator.js
